pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "codexwatch"
version = "0.1.0"
description = "Runner foundation for merged PR watch jobs."
requires-python = ">=3.12"
dependencies = [
  "openai>=1.60.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/codexwatch"]
